Updated the architecture of interaction with LLM models, added Health_Insurance dataset
