+++ examples/generate_description.py
@@ -1,12 +1,12 @@
-from fedotllm.data import Dataset
-from fedotllm.web_api import WebAssistant
-from fedotllm.actions import ModelAction
+from fedot_llm.data.data import Dataset
+from fedot_llm.web_api import WebAssistant
+from fedot_llm.language_models.actions import ModelAction
 from pprint import pprint
 
 
 def main():
     LLAMA8B = "http://10.32.2.2:8672/v1/chat/completions"
-    dataset = Dataset.load_from_path('datasets/titanic')
+    dataset = Dataset.load_from_path('datasets/Health_Insurance')
     train = dataset.splits[0]
     model = WebAssistant(LLAMA8B, model_type='8b')
     action = ModelAction(model=model)
