+++ fedot_llm/language_models/llms/web_model.py
@@ -1,0 +1,19 @@
+from typing import Dict, List, Optional
+from fedot_llm.language_models.base import BaseLLM
+import requests
+import json
+
+class CustomWebLLM(BaseLLM):
+    def __init__(self, url: str, model: str, timeout: Optional[int] = 10):
+        self.url = url
+        self.model = model
+        self.timeout = timeout
+    
+    def _generate(self, formatted_prompt: Dict[str, List[Dict[str, str]]]) -> str:
+        try:
+            response = requests.post(url=self.url, json=formatted_prompt, timeout=self.timeout)
+            response.raise_for_status()
+        except requests.exceptions.HTTPError as err:
+            raise RuntimeError(err)
+        return json.loads(response.text)['choices'][0]['message']['content']
+
